requirements.txt
fastapi==0.115.2
gunicorn==22.0.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
pydantic==2.9.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
email-validator==2.2.0
python-multipart==0.0.9
alembic==1.13.1
celery==5.4.0
redis==5.0.8
stripe==9.3.0
resend==2.2.0
google-auth==2.35.0
google-cloud-storage==2.18.2
pytest==8.3.2
